# Turn O(n^2) loop to O(n)