Currently inside `ido-vertical-completions`, the loop is using `nth` twice to
modify the list of completions, this results in O(n^2) time and slows down
showing the list of completions tremendously. This PR turns that loop to O(n) time.

Benchmark results with `(flx-ido-mode)`, `C-u 100 M-x benchmark RET (ido-vertical-completions "pdpid")` 

Before:
`Elapsed time: 5.769146s (0.345513s in 3 GCs)`

After:
`Elapsed time: 1.223311s (0.226972s in 2 GCs)`